requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
